feat: header 퍼센트바 분리 리렌더링 최적화
